feat: rename functions and files from being Modal focused to being focused on what part of the game we are at
