+++ redux/types.js
@@ -6,3 +6,4 @@ export const RESET = 'RESET'
 
 export const OPEN_QUESTION_MODAL = "OPEN_QUESTION_MODAL";
 export const CLOSE_QUESTION_MODAL = "CLOSE_QUESTION_MODAL";
+export const CREATE_TRIVIA_QUESTION = "CREATE_TRIVIA_QUESTION";
